+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Preferably use **Added**, **Changed**, **Removed** and **Fixed** topics in each 
 
 ## [Unreleased]
 
+## [1.3.3](https://github.com/quintoandar/butterfree/releases/tag/1.3.3)
+* Fix Cassandra Config and Numpy version ([#364](https://github.com/quintoandar/butterfree/pull/364))
+
+## [1.3.2](https://github.com/quintoandar/butterfree/releases/tag/1.3.2)
+* Fix publish script ([#362](https://github.com/quintoandar/butterfree/pull/362))
+
 ## [1.3.2](https://github.com/quintoandar/butterfree/releases/tag/1.3.2)
 * Fix publish script ([#360](https://github.com/quintoandar/butterfree/pull/362))